Refactoring to a udev rule

+++ RetroPie-Setup/scriptmodules/supplementary/supacec.sh
@@ -25,18 +25,32 @@ function sources_supacec() {
 }
 
 function remove_supacec() {
-    sudo update-rc.d -f supacec remove
-    sudo /etc/init.d/supacec stop
-    sudo rm -f /usr/bin/supacec 
-    sudo rm -f /etc/init.d/supacec
+
+    #lets compensate for our old setup
+    if [ -f /etc/init.d/supacec ]; then
+        sudo update-rc.d -f supacec remove
+        sudo /etc/init.d/supacec stop
+        sudo rm -f /etc/init.d/supacec
+    fi
+
+    if [ -f /usr/bin/supacec ]; then
+        sudo rm -f /usr/bin/supacec
+    fi
+    
+    # lets also remove our new setup
+    if [ -f /usr/bin/supacec ]; then
+        sudo rm -f /etc/udev/rules.d/10-cec.rules
+    fi
+    
+    sudo udevadm control --reload-rules   
+
 }
 
 function install_supacec() {
-    sudo cp "$md_build/supacec.d" /etc/init.d/supacec
-    sudo cp "$md_build/supacec" /usr/bin/supacec
-    sudo chmod +x /usr/bin/supacec
-    sudo chmod +x /etc/init.d/supacec
-    sudo update-rc.d supacec defaults
-    sudo /etc/init.d/supacec start
+    
+    remove_supacec   
+    sudo cp "$md_build/10-cec.rules" /etc/udev/rules.d/10-cec.rules    
+    sudo udevadm control --reload-rules
+
 
 }
